requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.7
django-betterforms==1.2
django-bootstrap-form==3.4
django-filter==2.4.0
django-stdimage==5.2.0
django-tempus-dominus==5.1.2.13
django-utils-six==2.0
django-widget-tweaks==1.4.8
gunicorn==20.0.4
Pillow==8.1.2
progressbar2==3.53.1
python-decouple==3.4
python-utils==2.5.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
